requirements.txt
streamlit
openai
beautifulsoup4
requests
python-dotenv
plotly
pandas
Pillow